refactor: Move Marquee and TweetCard components
