+++ src/app/components/TweetSection/tweets.d.ts
@@ -1,0 +1,6 @@
+export type TweetDataType = {
+  name: string;
+  username: string;
+  body: string;
+  img: string;
+};
