+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import FAQ from "@/app/components/FAQ/FAQ";
-import TweetSection from "@/components/Marquee/TweetSection";
+import TweetSection from "@/app/components/TweetSection/TweetSection";
 
 export default function Home() {
   return (
