+++ src/configs/tweetsData.ts
@@ -1,71 +1,73 @@
-export const firstMarqueetweetsData = [
+import { TweetDataType } from "@/app/components/TweetSection/tweets";
+
+export const firstMarqueetweetsData: TweetDataType[] = [
   {
     name: "صالح شجاعی",
     username: "@felxxbs",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/saleh-shojaei.png",
   },
   {
     name: "Fabrizio Fernandez",
     username: "@fab3304",
     body: "الان رسیدم محل برگزاری همایشحاجی ﺧﻴﻠﯽ ﺧﻔﻨﻪ 😍#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/Fabrizio-Fernandez.png",
   },
   {
     name: "صالح شجاعی",
     username: "@felxxbs",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/saleh-shojaei.png",
   },
   {
     name: "Fabrizio Fernandez",
     username: "@fab3304",
     body: "الان رسیدم محل برگزاری همایشحاجی ﺧﻴﻠﯽ ﺧﻔﻨﻪ 😍#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/Fabrizio-Fernandez.png",
   },
   {
     name: "صالح شجاعی",
     username: "@felxxbs",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/saleh-shojaei.png",
   },
   {
     name: "Fabrizio Fernandez",
     username: "@fab3304",
     body: "الان رسیدم محل برگزاری همایشحاجی ﺧﻴﻠﯽ ﺧﻔﻨﻪ 😍#frontchapter1402",
-    img: "https://avatar.vercel.sh/jack",
+    img: "/images/tweets/profile-picture/Fabrizio-Fernandez.png",
   },
 ];
 
-export const secondMarqueetweetsData = [
+export const secondMarqueetweetsData: TweetDataType[] = [
   {
     name: "Darius Flynn",
     username: "@flynnn",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jane",
+    img: "/images/tweets/profile-picture/Darius-Flynn.jpg",
   },
   {
     name: "Darius Flynn",
     username: "@flynnn",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jane",
+    img: "/images/tweets/profile-picture/Darius-Flynn.jpg",
   },
   {
     name: "Darius Flynn",
     username: "@flynnn",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jane",
+    img: "/images/tweets/profile-picture/Darius-Flynn.jpg",
   },
   {
     name: "Darius Flynn",
     username: "@flynnn",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jane",
+    img: "/images/tweets/profile-picture/Darius-Flynn.jpg",
   },
   {
     name: "Darius Flynn",
     username: "@flynnn",
     body: "رفقا می‌تونید با رزرو زودتر بلیط اقمتگاهتون، راحت‌تر توی همایش شرکت کنید#frontchapter1402",
-    img: "https://avatar.vercel.sh/jane",
+    img: "/images/tweets/profile-picture/Darius-Flynn.jpg",
   },
 ];
